requirements.txt
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
bleach==6.0.0
bokeh==3.1.1
boltons @ file:///croot/boltons_1677628692245/work
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
colorcet==3.0.1
conda==23.3.1
conda-build==3.24.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cubinlinker-cu11==0.3.0.post1
cuda-python==11.8.1
cudf-cu11==23.4.1
cuml-cu11==23.4.1
cupy-cuda11x==11.6.0
cycler==0.11.0
Cython==0.29.34
dask==2023.3.2
dask-cuda==23.4.0
dask-cudf-cu11==23.4.1
datashader==0.15.0
datashape==0.5.2
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distributed==2023.3.2.1
dnspython==2.3.0
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.1
exceptiongroup==1.1.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
expecttest==0.1.4
fastrlock==0.8.1
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.39.4
fsspec==2023.5.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
h5py==3.8.0
holoviews==1.16.2
huggingface-hub==0.15.1
hypothesis==6.75.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.1
importlib-metadata==6.6.0
iopath==0.1.10
ipython @ file:///croot/ipython_1680701871216/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kiwisolver==1.4.4
lazy_loader==0.2
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
linkify-it-py==2.0.2
lit==16.0.5
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.1
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdit-py-plugins==0.4.0
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
multipledispatch==0.6.0
mypy-extensions==1.0.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
nvtx==0.2.5
omegaconf==2.3.0
opencv-python-headless==4.8.0.76
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
panel==1.1.0
param==1.13.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.4.0
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pkginfo @ file:///croot/pkginfo_1679431160147/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
portalocker==2.7.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==4.21.12
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptxcompiler-cu11==0.7.0.post1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==10.0.1
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyct==0.5.0
Pygments @ file:///croot/pygments_1683671804183/work
pylibraft-cu11==23.4.1
pynndescent==0.5.10
pynvml==11.4.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyre-extensions==0.0.23
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz @ file:///croot/pytz_1671697431263/work
pyviz-comms==2.3.1
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1670514731622/work
qudida==0.0.4
raft-dask-cu11==23.4.1
rdkit==2023.3.3
requests @ file:///croot/requests_1682607517574/work
rmm-cu11==23.4.1
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.24.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post5
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1680518478486/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
submitit==1.4.5
sympy==1.12
tabulate==0.9.0
tblib==1.7.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.0+cu117
torchaudio==2.0.2
torchdata @ file:///__w/_temp/conda_build_env/conda-bld/torchdata_1682362130135/work
torchelastic==0.2.2
torchmetrics==0.10.3
torchtext==0.15.2
torchvision==0.15.0+cu117
tornado==6.3.2
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets @ file:///croot/traitlets_1671143879854/work
treelite==3.2.0
treelite-runtime==3.2.0
triton==2.0.0
types-dataclasses==0.6.6
typing-inspect==0.8.0
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
uc-micro-py==1.0.2
ucx-py-cu11==0.31.1
umap-learn==0.5.3
urllib3 @ file:///croot/urllib3_1680254681959/work
wandb==0.15.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
xarray==2023.5.0
xformers==0.0.18
xyzservices==2023.5.0
yacs==0.1.8
zict==3.0.0
zipp==3.15.0
zstandard @ file:///croot/zstandard_1677013143055/work
